create and use HTMLEdit class. compute style on mounted.
